backend/pyproject.toml
[project]
name = "automation-hub-backend"
version = "0.1.0"
description = "Automation Hub Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "cryptography>=42.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
    "anthropic>=0.40.0",
    "boto3>=1.34.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.9",
    "apscheduler>=3.10.4",
    "bcrypt>=4.1.0",
    "python-jose[cryptography]>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
